+++ ellie_engine/core/src/definite/types/double.rs
@@ -3,6 +3,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(PartialEq, Default, Debug, Clone, Serialize, Deserialize)]
 pub struct DoubleType {
-    pub value: f64,
+    pub value: f32,
     pub pos: defs::Cursor,
 }
